requirements-colab.txt
accelerate==1.2.1
bitsandbytes==0.45.0
datacollective>=0.5.0
datasets==3.2.0
edge-tts>=6.1.0
evaluate==0.4.3
fastapi==0.115.6
gTTS==2.5.4
gradio==5.12.0
jiwer==3.1.0
librosa==0.10.2
loguru==0.7.3
peft==0.14.0
pyloudnorm==0.1.1
pyyaml==6.0.2
python-multipart==0.0.20
sentencepiece==0.2.0
soundfile==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.48.1
uvicorn[standard]==0.34.0
